Relational subset representation of labels
This is a prototype of representing labels as sets. By contrast to the
ADT representation, sets do not have a structure, they are just ordinary
symbols. Set membership is represented with a relation that is true
when the element is in the set. Here, this is fieldHasLabelElt(f, lb)
where the set name is the same as the field and this is true if the
label element lb is an element of the label of f.

This gets around the issue of needing a normal form for labels like the
ADT version does. SubsetEq can be used for doing label checks, and this
is implmented in terms of NotSubsetEq.