Prototyping a conventional IFC labelcheck
This code prototypes a conventional IFC label check to determine if there are
information leaks in any of the modules in a graph. A relational representation
of the lattice over sets using subset as the set operator. To get around
the need for either computing join over a set of labels (in other words \bigcup join),
a relation "notLabelChecks" is used to find a single counter-example
among input-output pairs and a module label checks if it not
notLabelChecks.

It looks like this is also amenable to separating the
rules into a dataflow analysis pass for label inference and for the
check -- the inference would fill in "fieldHasLabel".

It should also be possible to support downgrading using two sets
of rules for notLabelCheck, but this also needs to be tried.