Attempt Labelchecking using set membership + meet/join
This attempt works by computing the meet/join over the labels of field
lists which are sets using an isMember relation.
It looks like this has the same issue as the approach where we look for
a counterexample at the "moduleChecks" level. The flowsTo check still
works by defning notFlowsTo which essentially finds a counterexample.
See the LABELCHECKING_STRATEGIES.md to see why the counterexample
approach seems not to work.