First draft of simplified Manifest fact structures.
This commit creates the Datalog data structures and relationships to be used in specifying input facts to the dataflow analysis. In many cases, these are simplified, mostly by omitting type information. The goal here is to be able to represent a simple manifest containing a few particle specs that only read and write handles, claim a label on some handle connection, and check it later.

This commit assumes that the manifest facts we provide to the dataflow analysis will have a structure that is matched closely to the original manifest protobuf. Eventually, we will be creating a tool that translates manifest protobufs to be analyzed into facts. Having a structure that matches the original protobuf as closely as possible in Datalog will allow this translation tool to be simple and insulate it from knowledge about the inner workings of the Datalog dataflow analysis.

This is in first-draft state; I am not certain that this even compiles at this point. Nonetheless, it is probably worth putting it up so others can see what I'm thinking at this point.