Add Authorization Logic (#1)
This commit adds a rust implementation of an authorization logic compiler. This is added as the result of a git squash, so most of the history is not preserved. In case the history becomes relevant, the history can be found in [the pull request that generated this commit](https://github.com/google-research/raksha/pull/1)