Add simple dataflow analysis and test harness for testing.
This commit adds a very simple dataflow analysis that just propagates
tags across edges in a naive fashion.

In addition, it adds a test infrastructure in the directory arcs_fact_tests.
This infrastructure expects tests to be provided as datalog scripts
that #include analyze_tag_checks_test_helper.dl and otherwise contain
only dataflow facts such as edges, claims, checks, and the expected
result of those checks. Each of these scripts is compiled into
a separate test binary which will exit with the success (0) exit code if
the datalog code produces no output and the 1 exit code otherwise.