Implement IFC + auth logic for downgrading
A check that does IFC and uses authorization logic for downgrading is
implemented. A low-level IR of particles is also given (see also
comments of flat_graph_ir for more info). Many predicates defined in
taint_with_auth_logic.dl are deliberately of the same name and type as
those given in taint.dl, so that other files can import either one to
use either analysis, potentially. An interface between authorization
logic and the IFC analysis is given.