Simplify the test infrastructure by removing helpers.
Gogul pointed out that many of the test helpers in previous commits
provided too much mechanism when we know that some pieces may change in
the future. This removes all of the helpers and replaces them with
a single testPasses output relation. Each test provides a rule for
populating this relation which indicates what must be done for the test to
pass.