Remove unused decls/rules for creating and testing checks.
After the simplification of the fact test infrastructure, the rules
relating to checks and evaluating whether or not they pass are no longer
needed. This commit gets rid of those defunct rules (and the entire
analyze_tag_checks.dl header). Because the test helper no longer centers
around these checks, rename it to fact_test_helper.