Switch testPasses -> testFails; allow multiple testFails rules.
Initially, I created tests using a relation called testPasses for two
reasons:

1. To allow tests to state what they want rather than the opposite of
   what they want.
2. To have a natural check that we included a test predicate at all.

@bgogul pointed out that we might want to add multiple rules stating the
requirements of the test and fail if any one of those rules failed. This
was not possible to do with testPasses. This commit switches to using
testFails, which naturally allows for expressing multiple requirements
in multiple rules.

This, however, opens us up to the possibility that someone might forget
to put a testFails at the bottom of a test entirely. I could not find
a good way to fix this in datalog, so as a very crude guardrail, I added a
shell test that just greps each file for testFails.