Rewrite the test infrastructure to rely on a new TEST_RULE macro. (#37)
Rewrite the test infrastructure to rely on a new TEST_CASE macro.

Add a new TEST_CASE macro that registers an aspect in allTests and
adds an aspect to testPasses if the attached condition succeeds. This
allows us to populate allTests with all tests even if not all tests succeed.
This allows us to correctly create a testFails relation that contains all
declared tests that did not pass.

TEST_CASE is created so that it takes the place of a test rule's head, and
takes as its argument the aspect it tests.