Split the Tag domain into Privacy, Integrity, and Claimed tags.
As discussions on Raksha have evolved, it has become clear that there is
no coherent notion of a simple, unqualified "tag". Privacy tags spread
like taint in a worst-case scenario, integrity tags disappear unless
properly handled, and the initial tag the user claims immediately splits
into each of the above kinds. It makes sense to ask whether a Privacy
tag "may" be present, but not whether it "must" be present; it does not
make sense to ask whether an Integrity tag "may" be present, but is
sensible to ask whether it "must" be present.

To address this reality, this commit introduces new base types for
Privacy, Integrity, and Claimed tags and places the appropriate type on
the appropriate "presence" queries.