Add the AccessPath, used to represent paths while converting to Datalog.
The AccessPath class represents a descent through the substructures of
an Arcs Manifest, and is used for representing such paths when drawing
edges and deciding which edges to draw (by, for instance, checking which
partial AccessPaths match each other in ParticleSpec connections).

The AccessPath is designed to be built bottom up. As such, the order of
components are reversed internally, allowing for pushing on parent
components as an access path is returned from leaf fields towards the
containing Recipe.