Use the .bazelrc to specify C++17 throughout raksha. (#67)
* Use the .bazelrc to specify C++17 throughout raksha.

This is a bit of a hack; the best way to specify the C++ standard across
the entire project is to make a custom CcToolchainConfigInfo, but that
is incredibly heavyweight. This allows us to set the C++ standard in one
place without tackling a gigantic configuration problem. This is very
important for importing Abseil, as there are issues if Bazel's magic
decides your system prefers C++14 (like it does on my machine) and then
tries to build Abseil with C++14 while the rest of your project uses
C++17 (see https://github.com/abseil/abseil-cpp/issues/218).