Focus on types and add Selector, making AccessPath SelectorAccessPath.
This change makes the AccessPath, previously just a reverse-order vector
of strings, have a bit more structure by making it a reverse-order
vector of Selectors. A Selector is any additional qualifier by which you
might descend through a type tree, such as a field access, index into an
array, tuple access, etc. For now, only FieldSelector, representing
a descent into a field, is implemented.

In addition, this change focuses the class on representing subpaths that
descend through the type tree, renaming it to SelectorAccessPath. We may
add an additional AccessPath structure that is an AccessPathRoot and
a SelectorAccessPath in the future.

Also, to prevent accidentally making copies, we make the copy
constructor private and require it to be explicitly invoked, while
making the move constructor public. Because we are returning these
AccessPaths in a bottom-up fashion up the type hierarchy, preventing
accidental copies helps prevent us from accidentally making performance
quadratic.