Add the AccessPathSelectors, used to represent paths through type members while converting to Datalog.      (#68)
* Add the AccessPath, used to represent paths while converting to Datalog.

The AccessPath class represents a descent through the substructures of
an Arcs Manifest, and is used for representing such paths when drawing
edges and deciding which edges to draw (by, for instance, checking which
partial AccessPaths match each other in ParticleSpec connections).

The AccessPath is designed to be built bottom up. As such, the order of
components are reversed internally, allowing for pushing on parent
components as an access path is returned from leaf fields towards the
containing Recipe.

* Make method names match Google C++ naming conventions.

* Move AccessPath to src/ir and namespace raksha::ir.

Address @bgogul's (offline) review comment that AccessPath should be
considered part of Raksha's IR. This change moves AccessPath to a new ir
subdirectory of src and puts it in namespace raksha::ir.

* Simplify AccessPath's operator== by using std::vector's operator==.

* Focus on types and add Selector, making AccessPath SelectorAccessPath.

This change makes the AccessPath, previously just a reverse-order vector
of strings, have a bit more structure by making it a reverse-order
vector of Selectors. A Selector is any additional qualifier by which you
might descend through a type tree, such as a field access, index into an
array, tuple access, etc. For now, only FieldSelector, representing
a descent into a field, is implemented.

In addition, this change focuses the class on representing subpaths that
descend through the type tree, renaming it to SelectorAccessPath. We may
add an additional AccessPath structure that is an AccessPathRoot and
a SelectorAccessPath in the future.

Also, to prevent accidentally making copies, we make the copy
constructor private and require it to be explicitly invoked, while
making the move constructor public. Because we are returning these
AccessPaths in a bottom-up fashion up the type hierarchy, preventing
accidental copies helps prevent us from accidentally making performance
quadratic.

* Add additional tests and comments for Selector classes.

* Address @bgogul's comments.

In particular:

Make some copy constructors private to prevent
accidentally copying in various places.

Fix a silent copy of std::string in the FieldSelector constructor.

Fix include path format and order.

Address some friction in tests around copy restrictions by adding
a Copyable wrapper, which makes an explicitly-copyable class into an
implicitly-copyable class.

Rename SelectorAccessPath -> AccessPathSelectors.

* Inline AccessPathSelector's operator== into its header file.

* Simplify AccessPathSelectors::ToString using a Formatter.

* Inline AccessPathSelectors::ToString into header.

Use a lambda in the place of a callable struct to call ToString on each
Selector when creating a string representation of an AccessPathSelectors
object. This allows the AccessPathSelectors::ToString method to be
simple enough to inline into the header and allows us to delete the .cc
file.

* Rename some locals from selector_access_path -> access_path_selectors.

* Remove explicit Copy methods, use default C++ copy constructors.

* Improve documentation for tests of selectors.

* Put FieldSelector::ToString all on one line.

* Make calls to MakeSelectorAccessPathFromString copy-elision friendly.

Make the incoming parameter no longer const and remove some unnecessary
local consts elsewhere.

* Refactor Selector class to inline remaining funcs in .cc file.

* Cleaned up BUILD file for ir package.