Add first few Type classes and a GetAccessPaths in each.
Add the PrimitiveType and EntityType classes and an interface Type
from which both inherit. Also add a method GetAccessPaths which produces
the SelectorAccessPathsSet describing the SelectorAccessPaths possible
through that type.

Also, make gtest.h pull in the gmock header, as it is required for more
advanced matchers, some of which we use in the tests generated for the
access path collecting functions.