Start of Combined Policy Example
This is a draft PR to check my understanding of how Arcs manifests work.
The code in VisionPipeline.arcs is meant to describe the same
graph as in
src/analysis/souffle/tests/taint_with_logic/test_taint_with_auth_logic.dl
while trying to use as few Arcs features as possible. Is it a complete
Arcs graph, though? Or is more needed