Add UnrootedTagClaim and TagClaim which claim a tag is on an path.
These two classes both represent a claim upon a particular path. The
difference is that the UnrootedTagClaim claims a tag is on a particular
AccessPathSelectors object, while TagClaim is a claim upon a full
AccessPath.

Why have two separate classes? The reason is that a ParticleSpec
contains claims that are rooted at the top of the ParticleSpec rather
than what Raksha considers the "true" root: the top of the Recipe. The
UnrootedAccessPath is the data that we derive from the ParticleSpec,
which is then instantiated at the various particles by associating it
with a root string. The resulting TagClaim is a fully-formed Datalog
fact, ready to be printed out.