Add a structured AccessPath root, which may be Spec or Concrete.
This is used to express what was previously called the "rooted" and
"unrooted" case: an AccessPath that is rooted at a ParticleSpec has
a SpecAccessPathRoot, while one rooted at an instantiated structure has
a ConcreteAccessPathRoot. Attempting to get a Datalog string for
a SpecAccessPathRoot will cause an assertion failure. Similarly, add an
Instantiate function to AccessPath and TagClaim to allow instantiating
those classes with a ConcreteAccessPathRoot if they previously had
a SpecAccessPathRoot. Get rid of the RootInstantiated class.