Keep AccessPath info about ParticleSpec roots, add structure to roots.
Previously, we represented root classes as just empty spec roots and
concrete roots which carry a single string. This threw away too much
information, and would make it impossible to associate a ParticleSpec's
HandleConnectionSpec with the handle it was supposed to connect to. This
change preserves that information from the protos and adds additional
structure to the AccessPathRoot classes, allowing for preserving the two
types of roots we know we will have to deal with for the MVP:
HandleConnectionSpecAccessPathRoot and HandleConnectionAccessPathRoot.