Add TagClaim class; add structured representation of AccessPathRoot. (#95)
* Add UnrootedTagClaim and TagClaim which claim a tag is on an path.

These two classes both represent a claim upon a particular path. The
difference is that the UnrootedTagClaim claims a tag is on a particular
AccessPathSelectors object, while TagClaim is a claim upon a full
AccessPath.

Why have two separate classes? The reason is that a ParticleSpec
contains claims that are rooted at the top of the ParticleSpec rather
than what Raksha considers the "true" root: the top of the Recipe. The
UnrootedAccessPath is the data that we derive from the ParticleSpec,
which is then instantiated at the various particles by associating it
with a root string. The resulting TagClaim is a fully-formed Datalog
fact, ready to be printed out.

* Add RootInstantiated, which associates any unrooted class with a root.

By creating a RootInstantiated class, we can wrap any class initially
constructed without a root with additional root information. This allows
adding root information after the fact without mutating the original
rootless class. We remove TagClaim, leaving only UnrootedTagClaim, as
TagClaim can be replaced with RootInstantiated<UnrootedTagClaim>.

* Add a structured AccessPath root, which may be Spec or Concrete.

This is used to express what was previously called the "rooted" and
"unrooted" case: an AccessPath that is rooted at a ParticleSpec has
a SpecAccessPathRoot, while one rooted at an instantiated structure has
a ConcreteAccessPathRoot. Attempting to get a Datalog string for
a SpecAccessPathRoot will cause an assertion failure. Similarly, add an
Instantiate function to AccessPath and TagClaim to allow instantiating
those classes with a ConcreteAccessPathRoot if they previously had
a SpecAccessPathRoot. Get rid of the RootInstantiated class.

* Refactor code slightly by giving AccessPath its own CreateFromProto.

Previously, we had not done this because we were treating an "unrooted"
tag claim as if it didn't have a real AccessPath on it. Now that we
allow for a SpecAccessPathRoot, it makes more sense to parse an
AccessPath from an AccessPathProto.

* Fix typo.

* Make Instantiate a member method rather than a static method.

This is more readable at the use sites.

* Keep AccessPath info about ParticleSpec roots, add structure to roots.

Previously, we represented root classes as just empty spec roots and
concrete roots which carry a single string. This threw away too much
information, and would make it impossible to associate a ParticleSpec's
HandleConnectionSpec with the handle it was supposed to connect to. This
change preserves that information from the protos and adds additional
structure to the AccessPathRoot classes, allowing for preserving the two
types of roots we know we will have to deal with for the MVP:
HandleConnectionSpecAccessPathRoot and HandleConnectionAccessPathRoot.

* Update some stale comments.

* Remove redundant whitespace.