Add TagCheck class and add class for sharing code btw Claim and Check. (#96)
* Add TagCheck class and add class for sharing code btw Claim and Check.

This commit introduces the TagCheck class. This class will eventually
represent a check of some arbitrary binary predicate involving tags
upon some AccessPath. For now, though, it represents a check that
a particular tag is present on an AccessPath.

Despite being semantically very different, the internal representation
of a TagClaim and a TagCheck are actually very similar: both annotate
a tag onto an AccessPath. The difference is merely that the claim
introduces the tag while the check asserts the tag. This similarity in
representation, however, allows us to use the new
TagAnnotationOnAccessPath to share code between the two classes.

* Remove redundant whitespace.