Auth logic side + fix interface
This commit includes the actual frontend for the authorization logic
code. The output is at raksha_examples/VisionPipelineAuthLogic.dl is
also checked in for now (though it should be removed eventually). The
output for VisionPipelineAuthLogic.dl does not use types (everything is
a symbol) whereas types are used throught the Raksha implementation, so
we need to bridge this gap somehow so that they can be integrated.