Add type information and GetAccessPaths to HandleConnectionSpec. (#100)
This PR adds type information to HandleConnectionSpec, preserving the
information that was already in the manifest.proto. Initially, I thought
I wouldn't need this because my quick-and-dirty partial implementation
of the proto transformer looked like it did not need it. However, that
was more top-down than the cleaner approach has ended up being, and the
type information was necessary for building the ParticleSpec.

This PR also adds a method GetAccessPaths to HandleConnectionSpec, which
returns a vector containing all of the full AccessPaths that can be
accessed through the HandleConnectionSpec and end in a leaf field of the
type on HandleConnectionSpec. This is useful for building the edges
within a ParticleSpec.

This PR also adds operator== to the AccessPathRoots and some additional
field accessors to AccessPath. operator== on roots will eventually
become important for instantiating particle specs and was useful for
testing HandleConnectionSpec's new methods, so