Add DatalogFacts class, which gathers facts from Manifest proto.
When the DatalogFacts class is constructed from a ManifestProto, it
contains all of the objects that describe the facts needed by the
dataflow graph that are implied by that ManifestProto. DatalogFacts also
contains a ToString method which prints all of the facts out to
a string.

Note that this is pretty quick and dirty to move towards an MVP; we
should clean this up as described in issue #107.