Combined Policy Example (#92)
This is an end-to-end example that combines arcs manifests, taint analysis, and policy in authorization logic. At the moment, this code is useful for figuring out the loose ends for integration, but it does not yet do much to demonstrate the usefulness of adding authorization logic.

Explanation of files:
third_party/arcs/examples/
VisionPipeline.arcs --- input arcs manifests

src/analysis/souffle/
VisionPipelineOut.dl --- imagined output of converting arcs manifest into datalog (makes use of flat_graph_ir.dl but that's not the only way)
VisionPipelineAuthLogic.dl --- approximate output from auth logic (ignoring some wrinkles about types that need to be dealt with)
VisionPipelineTagOwners.dl --- only other piece of policy code aside from the claims/checks and what comes from auth logic

rust/tools/authorization-logic/raksha_examples/
VisionPipelineAuthLogic -- this is the input file from authorization logic
VisionPipelineAuthLogic.dl -- this is the actual output from auth logic. This should not remain in the repository but it is included for now just to illustrate a small wrinkle with types that needs to be resolved for integration.

Related to #80