Draw edges between Handles and HandleConnections only when needed.
Prior to this commit, we were always drawing edges between
a HandleConnection and the Handle to which it connected in
a bidirectional fashion, even when the HandleConnection only read or
wrote. The graph produced this way was slightly easier to make and was
still correct, as the extra edges were ignored by the edges drawn within
a ParticleSpec. The argument for this being correct, however, was not at
all intuitive. This commit makes the Datalog fact generator only draw
edges between Handles and HandleConnections in the direction implied by
the read/write behavior of the HandleConnectionSpec.