Add DatalogFacts class, which gathers facts from Manifest proto. (#108)
* Add DatalogFacts class, which gathers facts from Manifest proto.

When the DatalogFacts class is constructed from a ManifestProto, it
contains all of the objects that describe the facts needed by the
dataflow graph that are implied by that ManifestProto. DatalogFacts also
contains a ToString method which prints all of the facts out to
a string.

Note that this is pretty quick and dirty to move towards an MVP; we
should clean this up as described in issue #107.

* Fix duplicate hash implementation due to merge mistake.

* Improve comments and formatting in CreateFromManifestProto.

* Fix a grammatical error in a comment.

* Fix accessor method types and put on single line for readability.

* Remove a premature std::move call.

* Make large Manifest input string into raw string for readability.

* Remove some unused AccessPath globals in DatalogFacts test.

An earlier version of the DatalogFacts test had more cases in which
a field2 field was defined. These were removed to reduce the size of the
test. Remove these extraneous AccessPaths that referred to those cases.

* Use raw strings to reduce use of escape characters.

* Draw edges between Handles and HandleConnections only when needed.

Prior to this commit, we were always drawing edges between
a HandleConnection and the Handle to which it connected in
a bidirectional fashion, even when the HandleConnection only read or
wrote. The graph produced this way was slightly easier to make and was
still correct, as the extra edges were ignored by the edges drawn within
a ParticleSpec. The argument for this being correct, however, was not at
all intuitive. This commit makes the Datalog fact generator only draw
edges between Handles and HandleConnections in the direction implied by
the read/write behavior of the HandleConnectionSpec.

* State expected outputs in datalog_facts_test as raw strings.

This commit gets rid of the various intermediate constants and helpers
and just states the expected outputs as a simple collection of raw
strings.

* Break up big DatalogFacts test into a test fixture and subtests.