Remove newline at end of fact Datalog strings.
The initial ToString functions for classes representing Datalog facts
added trailing newlines for those facts. This was a mistake, as this
made the expected outputs in tests testing these functions much uglier.
This PR fixes this issue by moving the newlines to the delimiter in the
StrJoin function combining the datalog facts instead. This PR
fixes #110.