Add new DerivesFromClaim, alter edge drawing in ParticleSpec to use it.
The DerivesFromClaim claims that, within some ParticleSpec, an output
AccessPath derives from an input AccessPath. A DerivesFromClaim both
acts as an explicitly-specified dataflow edge and a request to exclude
the target AccessPath from participating in the default process that
draws all input handle connections to all output handle connections.

This commit also adds a hash function to AccessPath (which was not
needed before) and tests it.

In addition, because we now have two different kinds of claim, this
renames claim fields to tag_claim in all cases except those in
DatalogFacts. We refrained from doing this there to avoid disrupting
@bgogul's work on an overarching binary to produce datalog facts.