Make claimHasTag indicate the particle making the claim.
In order to connect the claims made in the Arcs manifest files with the
principals reasoned about by the authorization logic, this commit adds
a new principal field to the claimHasTag relation, populated with the
name of the ParticleSpec making the claim. Assuming that ParticleSpecs
connect to a principal in the authorization logic in some way (for
instance, the organization that wrote the ParticleSpec is associated
with the ParticleSpec name), this allows us to bridge the gap between
the dataflow graph and the authorization logic with a simple interface.

This commit also updates the xform_to_datalog code that generates
claimHasTag datalog facts to produce the new form correctly and updates
examples using claimHasTag to put in best guesses for what the
particles in the claims should be.