Create TagClaims for full range of Arcs tag claims.
Previously, our parsing of Manifest Assume protos could only handle
those claims that were simple introductions of a single tag. With this
commit, we now handle the full range of 'and' and 'not' predicates that
can occur in an Arcs tag claim.

Note that this uses the saysDowngrades relation for expressing negative
tag claims. I wonder whether this conflicts with @aferr's plan for this;
I can add a separate relation if that would be better.

This fixes #137.