Make @aferr's derives from solution more precise using edge midpoints.
In PR #145, @aferr introduced a solution for dealing with derives from
without altering the dataflow graph by having the analysis generate
downgrades requests for all tags on that edge. This solution is quite
elegant, but his implementation had a flaw of clearing all tags on the
source of the edge, which may potentially overclear the taint tags.

This solution attempts to remedy that problem by adding a "midpoint"
AccessPath to every edge. This allows us to express a downgrades request
only for the tags on a particular edge by downgrading the tags on the
midpoint.

In addition, this PR adds some mechanism to turn on and off the
fact_test_helper convenience rule that causes all principals to own all
tags. Somewhat hackily, the test for the derives from/claimNotEdge
functionality tests the "derives from claim is trusted" vs. "derives
from claim is not trusted" by flipping this option on and off. We should
have a more principled way of expressing test variants (see #146).