Use two taint passes to permit stratification
This extends the approach of bgogul, but is stratifiable.

There is a first pass of the taint analysis called mayHaveTagNoDown that
populates the set of tags when there are no downgrades. This is used by
the saysDowngrades rule that implements claimNotEdge to determine
which tags are on the source of the edge. The downgrade requests in
saysDowngrades are used by the second taint analysis mayHaveTag which
does account for downgrades. Because mayHaveTag is not directly
implemented in terms of mayHaveTagNoDown, it should prevent tags from
being propagated further in the graph.

Because mayHaveTagNoDown overapproximates mayHaveTag, I think there is a
situation where tags that have already been downgraded away earlier in a
a sequence of particles are requested to be downgraded again even though
they are not there. But, this overapproximation is still _safe_ because
for the tags to be removed, these requests still need to be permitted by
the data owners and a request to remove an edge is a request to remove
ALL tags from the source.

This gives an example of the overapproximation. Imagine the graph:

---h1-->[A]---h2-->[B]---h3-->

where {h1,h2,h3} are handles and [A] [B] are particles.

if there are claims to remove BOTH edges (h1,h2) and (h2,h3), this will
generate a request to remove all of the tags on h2 from h3. This will
request removal of the tags that were orignally on h1 from h3
even though they were already removed by the claimNotEdge(h1, h2).
If we imagine there is also a second input to [B], called h4 with the
same tags as h1, the tags will also be requested for removal from h3
even though h4 should have added them back in. For these requests to
succeed, the data owner would still have to give permission for this
downgrade at [B], which might be unusual. So perhaps this loss of
precision in what is requested to be downgraded
is acceptable since it does not violate safety?

Co-authored-by: @bgogul
Co-authored-by: @markww