Make Cargo Test Great Again
Just prior to this commit cargo test would fail. It looks like this is
the commit where the regression began:
[PR 148](https://github.com/google-research/raksha/pull/148).
Because of reasons that are a mystery to me, it looks like
when this is called from C, the doc comments are interpreted as rust
code, causing syntax errors to be thrown because these comments are not
rust code. As a kludgy workaround, this commit gets the tests to pass again
by just adding some extra C-style comments. No one will have to know.