Make Cargo Test Great Again (#154)
Just prior to this commit cargo test would fail. It looks like this is
the commit where the regression began:
PR 148.
Rustdoc executes code examples, causing syntax errors to be thrown because these comments are not
rust code. This fixes that by making rustdoc aware that this code is actually datalog.