Refactor facts to instantiate with pointers rather than copies.
Previously, we have instantiated facts such as checks, claims, and edges
by creating a copy of the original fact with a new AccessPathRoot
indicating the instantiated root. This was always dubious from
a performance standpoint. However, as I started another PR in which
I was introducing we have started introducing complex checks with
predicate trees pointed to by unique_ptrs, the memory ownership
considerations became too complex to keep the status quo.

This commit instead introduces a wrapper for facts that may not be
instantiated called InstanceFact. InstanceFact contains a variant that
associates a pointer to the original fact with an Instantiator, which
contains the information needed to replace the uninstantiated root with
an instantiated root. There is another variant for facts that were
instantiated initially that just wraps that fact and implicitly
associates it with a NoopInstantiator (which just asserts that the roots
it encounters are already instantiated).

Rather than having an instantiated copy, Instantiators allow for
instantiating AccessPathRoots as they are encountered, preventing a copy
from being necessary.

The use of pointers rather than copies introduced lifetime
considerations in a number of places where they were not present before,
most notably in ManifestDatalogFacts, which now must store a list of
unique_ptrs to ParticleSpecs and Instantiators that it relies upon being
live at least as long as its component InstanceFacts.