Add predicate trees; check that expected output strs are valid Datalog.
This commit adds the Predicate class and its various subclasses, which
allow expressing complex predicates upon the presence of tags. It also
adds a ToDatalogRuleBody method for this class which prints out the
predicate as the rule body for a check.

The expected output strings for these Predicates are pretty complex and
somewhat difficult to make into valid Souffle datalog, so this commit
also adds a tool that will, given a vector of Datalog strings provided
to it via linking magic, print those strings out to a datalog file and
run Souffle on them to ensure that they are valid datalog. Although it
currently only works on rule bodies, it could easily be extended for any
kind of snippet of Datalog.