Use edge midpoints to make @aferr's derives from via downgrades more precise. (#147)
* Support safe removal of edges

This rule is meant to support the functionality of removing edges from
manifests:
     - that clearly does not introduce an unsafe escape hatch that could
     break policies (because it does not introduce downgrades that would not
     otherwise be possible).
     - without data owners needing to know anything about the application
     structure
     - without app behavior specification writers needing to know anything
     about the taint analysis or the tags that might appear on a handle

claimNotEdge(p, src, tgt) is an _attempt_ to remove an edge from the
analysis (by principal p)
saysDowngrades(p, path, tgt) is an _attempt_ to remove just one tag
from a path (by principal p).

The rule for downgrades(p, t) in authorization_logic.dl describes when
a downgrade is "safe" and the downgrade really happens (by consulting the
owners).

Because the analysis propagates the tags on source edges to dest edges, it is safe
to remove an edge when ALL of the tags on the source edge can be downgraded.
Therefore, a request to remove an edge is a request to remove all tags from
the source path. This rule populates a request to downgrade each tag on the
source path.
     - it is safe because these requests are only accepted if they would
     normally be by the downgrades(_, _) rule
     - it does not require the manifest writer to know anything about the
     tags in the analysis
     - data owners do not need to know about paths, and they will grant
     permission for edge removal indirectly by granting permission to remove
     tags (abstract representations of their data).

* Make @aferr's derives from solution more precise using edge midpoints.

In PR #145, @aferr introduced a solution for dealing with derives from
without altering the dataflow graph by having the analysis generate
downgrades requests for all tags on that edge. This solution is quite
elegant, but his implementation had a flaw of clearing all tags on the
source of the edge, which may potentially overclear the taint tags.

This solution attempts to remedy that problem by adding a "midpoint"
AccessPath to every edge. This allows us to express a downgrades request
only for the tags on a particular edge by downgrading the tags on the
midpoint.

In addition, this PR adds some mechanism to turn on and off the
fact_test_helper convenience rule that causes all principals to own all
tags. Somewhat hackily, the test for the derives from/claimNotEdge
functionality tests the "derives from claim is trusted" vs. "derives
from claim is not trusted" by flipping this option on and off. We should
have a more principled way of expressing test variants (see #146).

* Generate midpoint edges only for edges indicated by a claimNotEdge.

* Rename innerEdge -> resolvedEdge.

* Make any Souffle library where all principals own all tags testonly.

Co-authored-by: Andrew Ferraiuolo <aferr@google.com>