Add predicate trees; check that expected output strs are valid Datalog. (#157)
* Add predicate trees; check that expected output strs are valid Datalog.

This commit adds the Predicate class and its various subclasses, which
allow expressing complex predicates upon the presence of tags. It also
adds a ToDatalogRuleBody method for this class which prints out the
predicate as the rule body for a check.

The expected output strings for these Predicates are pretty complex and
somewhat difficult to make into valid Souffle datalog, so this commit
also adds a tool that will, given a vector of Datalog strings provided
to it via linking magic, print those strings out to a datalog file and
run Souffle on them to ensure that they are valid datalog. Although it
currently only works on rule bodies, it could easily be extended for any
kind of snippet of Datalog.

* Place GatherDatalogStrings in a namespace to prevent collisions.

* Fold predicates into general ir directory and namespace.

* Make DatalogString return a string_view instead of a string copy.

* Make predicate testdata testonly; make required changes in macros.

* Add comments explaining textproto for predicate test cases.

* Eliminate unnecessary local variables and the moves they require.

* Improve readability of test fixture constructor.

* Fix comment.

* Get rid of ir::predicates namespace, replace with ir.

* Fix comment on Predicate proto decoder.

* Tick-quote predicate class names in error messages.

* Remove unnecessary namespace alias.

* Added TODO comment requested in review.