Make will operate on handles
Previously both will and may would operate on tags, but it makes more
sense for will to operate on handles. By allowing the "will" relation to
work on handles, the app behavior spec writer does not need to know
anything about the taint analysis, and can describe usgae in terms of
handles. This also seems more likely to prevent errors since the spec
writer does not have the option of forgetting to mention some tags in
the usage.