Fix the sorting of declartion that was lost in a merge of PR #156. (#166)
This PR doesn't have any rust tests, but tests it indirectly through the C interface tests. However, the C interface tests only check that the output is always sorted. It does not create an input with unordered decls and make sure they are sorted. We should add unit tests to souffle emitter to capture this sorting behavior.